ToStringPartial -> To.ToStringAbstract
