+++ src/methods/to.js
@@ -11,6 +11,7 @@
 
 import type { Descriptor, CallableObjectValue } from "../types.js";
 import type { Realm } from "../realm.js";
+import { TypesDomain, ValuesDomain } from "../domains/index.js";
 import { GetMethod, Get } from "./get.js";
 import { Create } from "../singletons.js";
 import { HasProperty } from "./has.js";
@@ -34,6 +35,7 @@ import {
   Value,
 } from "../values/index.js";
 import invariant from "../invariant.js";
+import * as t from "babel-types";
 
 type ElementConvType = {
   Int8: (Realm, numberOrValue) => number,
@@ -742,6 +744,32 @@ export class ToImplementation {
       throw realm.createErrorThrowCompletion(realm.intrinsics.TypeError, "unknown value type, can't coerce to string");
     }
     return new StringValue(realm, str);
+  }
+
+  ToStringAbstract(realm: Realm, value: AbstractValue): AbstractValue {
+    if (value.mightNotBeString()) {
+      // If the property is not a string we need to coerce it.
+      let coerceToString = ([p]) => t.binaryExpression("+", t.stringLiteral(""), p);
+      let result;
+      if (value.mightNotBeNumber() && !value.isSimpleObject()) {
+        // If this might be a non-simple object, we need to coerce this at a
+        // temporal point since it can have side-effects.
+        // We can't rely on comparison to do it later, even if
+        // it is non-strict comparison since we'll do multiple
+        // comparisons. So we have to be explicit about when this
+        // happens.
+        result = realm.evaluateWithPossibleThrowCompletion(
+          () => AbstractValue.createTemporalFromBuildFunction(realm, StringValue, [value], coerceToString),
+          TypesDomain.topVal,
+          ValuesDomain.topVal
+        );
+      } else {
+        result = AbstractValue.createFromBuildFunction(realm, StringValue, [value], coerceToString);
+      }
+      invariant(result instanceof AbstractValue);
+      return result;
+    }
+    return value;
   }
 
   // ECMA262 7.1.2
